Random and scope

+++ functions/calculator/Calculator6.py
@@ -18,15 +18,17 @@ operations={
 
 def calculator():
   print (logo)
-  num1= int (input("Enter the first number:\t"))
+  num1= float (input("Enter the first number:\t"))
   for key in operations:
    print(f"{key}")
   to_continue=True
   while to_continue :
       operation_symbol=input("What is the operation you would like to perform?\n")
-      num2= int (input(f"Enter another number to  {operations[operation_symbol].__name__} with {num1}: \t"))
+      num2= float (input(f"Enter another number to  {operations[operation_symbol].__name__} with {num1}: \t"))
       calculation_function=operations[operation_symbol]
       answer=calculation_function(num1,num2)
+     
+     
       print(f"{num1} {operation_symbol} {num2} = {answer}\n")
       test=input (f"Type Y to continue operation with {answer} and N to start to a new operation \n")
       if(test=='Y'):
